workspace_assistant/requirements.txt
# Google ADK
google-adk>=0.1.0

# MCP client library
mcp>=1.0.0

# Google API clients
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0

# Environment
python-dotenv>=1.0.0

# CLI
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
